Add a new utils module for seeding
